building contacts and communication events
In this next phase of the project, I began implementing the Contact class and the communication events that belong to each contact. I also created entities for these classes via hibernate so they can be saved to a DB (MySQL). Added a controller for the contacts that is a Java object that can  search for and view Contacts stored in the database.